+++ CHANGES.md
@@ -34,6 +34,9 @@ cmdliner.
 
 ### Doc specification improvements and fixes
 
+- Add `?envs` optional argument to `Term.info`. Documents environment
+  variables that influence a term's evaluation and automatically
+  integrate them in the manual.
 - Add `Manpage.s_*` constants for standard manpage section names.
 - Add a `` `Blocks`` case to `Manpage.blocks` to allow block splicing
   (#69).  This avoids having to concatenate block lists at the